Revert "ecpriss: Fixed counters issue at OXTOR and eCPRI_CORE"
This reverts commit 5906de5cf19a97012b8172d75590d78e70d0c366.

+++ drivers/ecpri/ecpri_core/src/ecpriss_xbar/ecpriss_xbar.c
@@ -8,6 +8,7 @@
 #define ENABLE_BIT 1
 #define DISABLE_BIT 0
 #include "ecpriss_log.h"
+
 int xbar_irq_mapping = 0;
 
 void ecpriss_xbar_config_stats_update(void){
@@ -124,7 +125,7 @@ void ecpriss_xbar_config_stats_update_v2(void){
 
 void ecpriss_xbar_stats_update(void)
 {
-	uint32_t val=0;
+	uint64_t val=0;
 	int link_index;
 	uint64_t curr_wm_index = 0;
 
@@ -313,7 +314,7 @@ void ecpriss_xbar_stats_update(void)
 
 void ecpriss_xbar_stats_update_v2(void)
 {
-	uint32_t val=0;
+	uint64_t val=0;
 	int link_index;
 	uint64_t curr_wm_index = 0;
